--- app/src/main/java/com/numero/material_gallery/settings/SettingsActivity.kt
@@ -1,62 +1,0 @@
-package com.numero.material_gallery.settings
-
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import android.view.MenuItem
-import androidx.appcompat.app.AppCompatActivity
-import com.google.android.play.core.appupdate.AppUpdateInfo
-import com.google.android.play.core.appupdate.AppUpdateManager
-import com.google.android.play.core.appupdate.AppUpdateManagerFactory
-import com.google.android.play.core.install.model.AppUpdateType
-import com.google.android.play.core.install.model.UpdateAvailability
-import com.numero.material_gallery.R
-import kotlinx.android.synthetic.main.activity_settings.*
-
-class SettingsActivity : AppCompatActivity(R.layout.activity_settings), SettingsFragment.SettingsFragmentListener {
-
-    private lateinit var appUpdateManager: AppUpdateManager
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setSupportActionBar(toolbar)
-
-        supportActionBar?.apply {
-            setDisplayHomeAsUpEnabled(true)
-        }
-
-        appUpdateManager = AppUpdateManagerFactory.create(this)
-
-        supportFragmentManager.beginTransaction().replace(R.id.container, SettingsFragment.newInstance()).commit()
-    }
-
-    override fun onResume() {
-        super.onResume()
-        appUpdateManager.appUpdateInfo
-                .addOnSuccessListener { appUpdateInfo ->
-                    if (appUpdateInfo.updateAvailability() == UpdateAvailability.DEVELOPER_TRIGGERED_UPDATE_IN_PROGRESS) {
-                        doUpdate(appUpdateInfo)
-                    }
-                }
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return when (item.itemId) {
-            android.R.id.home -> {
-                finish()
-                true
-            }
-            else -> super.onOptionsItemSelected(item)
-        }
-    }
-
-    override fun doUpdate(appUpdateInfo: AppUpdateInfo) {
-        appUpdateManager.startUpdateFlowForResult(appUpdateInfo, AppUpdateType.IMMEDIATE, this, UPDATE_REQUEST_CODE)
-    }
-
-    companion object {
-        private const val UPDATE_REQUEST_CODE = 1
-
-        fun createIntent(context: Context): Intent = Intent(context, SettingsActivity::class.java)
-    }
-}
